+++ cascade/CHANGELOG.md
@@ -8,8 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
-* Add a `c:Cascade.Template.post_generate/1` callback to the `Cascade.Template`
-behaviour.
+* Add a `c:Cascade.Template.pre_generate/2` callback to the `Cascade.Template` for
+optional setup before template generation.
+* Add a `c:Cascade.Template.post_generate/1` callback to the `Cascade.Template` for
+post processing generated templates.
 
 ## [v0.1.0](https://github.com/sportradar/elixir-workspace/tree/cascade/v0.1.0) (2024-05-13)